# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OTO-RAPORT API - Ministry Compliance & Developer Management",
    "version": "1.0.0",
    "description": "REST API for OTO-RAPORT SaaS platform. Provides endpoints for Ministry Schema 1.13 compliance validation, developer profile management, and property data validation.\n\n**Authentication**: All endpoints require Supabase Auth session (JWT token in Authorization header).\n\n**Ministry Compliance**: Based on Polish law 'ustawa z dnia 21 maja 2025 r. o jawności cen mieszkań'.",
    "contact": {
      "name": "OTO-RAPORT Support",
      "email": "support@oto-raport.pl"
    }
  },
  "servers": [
    {
      "url": "https://otoraport-v2.vercel.app/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/developers/update": {
      "patch": {
        "tags": ["Developers"],
        "summary": "Update developer profile",
        "description": "Updates developer profile with Ministry Schema 1.13 fields. All 28 developer fields can be updated. Returns updated profile with validation status.",
        "operationId": "updateDeveloper",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeveloperUpdateRequest"
              },
              "examples": {
                "basicUpdate": {
                  "summary": "Update basic company info",
                  "value": {
                    "updates": {
                      "company_name": "Deweloper XYZ Sp. z o.o.",
                      "nip": "1234567890",
                      "phone": "+48 123 456 789",
                      "email": "kontakt@deweloper.pl"
                    }
                  }
                },
                "fullUpdate": {
                  "summary": "Update all sections",
                  "value": {
                    "updates": {
                      "company_name": "Deweloper XYZ Sp. z o.o.",
                      "legal_form": "Spółka z ograniczoną odpowiedzialnością",
                      "nip": "1234567890",
                      "headquarters_city": "Warszawa",
                      "headquarters_street": "Marszałkowska",
                      "headquarters_building_number": "100",
                      "headquarters_postal_code": "00-001",
                      "sales_office_city": "Warszawa",
                      "contact_method": "telefon, email, formularz"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Developer profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeveloperUpdateResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Conflict - duplicate data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Duplicate value - this data already exists"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "SupabaseAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeveloperUpdateRequest": {
        "type": "object",
        "required": ["updates"],
        "properties": {
          "developerId": {
            "type": "string",
            "format": "uuid",
            "description": "Optional developer ID (defaults to authenticated user's developer)"
          },
          "updates": {
            "type": "object",
            "description": "Fields to update (partial updates supported)",
            "properties": {
              "company_name": {
                "type": "string",
                "minLength": 1,
                "maxLength": 255,
                "example": "Deweloper XYZ Sp. z o.o."
              },
              "legal_form": {
                "type": "string",
                "maxLength": 100,
                "example": "Spółka z ograniczoną odpowiedzialnością"
              },
              "nip": {
                "type": "string",
                "pattern": "^\\d{10}$",
                "example": "1234567890"
              },
              "regon": {
                "type": "string",
                "pattern": "^\\d{9}$|^\\d{14}$",
                "example": "123456789"
              },
              "phone": {
                "type": "string",
                "maxLength": 50,
                "example": "+48 123 456 789"
              },
              "email": {
                "type": "string",
                "format": "email",
                "example": "kontakt@deweloper.pl"
              },
              "headquarters_city": {
                "type": "string",
                "maxLength": 100,
                "example": "Warszawa"
              },
              "headquarters_postal_code": {
                "type": "string",
                "pattern": "^\\d{2}-\\d{3}$",
                "example": "00-001"
              }
            }
          }
        }
      },
      "DeveloperUpdateResponse": {
        "type": "object",
        "required": ["success", "data"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "developer": {
                "type": "object",
                "description": "Updated developer profile (all fields)"
              },
              "validationStatus": {
                "type": "object",
                "properties": {
                  "complianceScore": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 100,
                    "example": 80
                  },
                  "missingFields": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["regon", "headquarters_street"]
                  },
                  "completionPercentage": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 100,
                    "example": 75
                  }
                }
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Internal server error"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request - validation errors",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Validation failed",
              "details": [
                {
                  "field": "updates.nip",
                  "message": "NIP musi składać się z 10 cyfr"
                }
              ]
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - missing or invalid authentication",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Unauthorized"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Forbidden - not your developer profile"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not Found - resource does not exist",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Developer not found"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Internal server error"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "SupabaseAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token obtained from authentication. Include in Authorization header: `Bearer {token}`"
      }
    }
  }
}